Move dataset pydantic models to server.schemas.datasets module
