+++ src/argilla/server/apis/v0/models/text_classification.py
@@ -25,8 +25,8 @@ from argilla.server.apis.v0.models.commons.model import (
     ScoreRange,
     SortableField,
 )
-from argilla.server.apis.v0.models.datasets import UpdateDatasetRequest
 from argilla.server.commons.models import PredictionStatus
+from argilla.server.schemas.datasets import UpdateDatasetRequest
 from argilla.server.services.search.model import (
     ServiceBaseRecordsQuery,
     ServiceBaseSearchResultsAggregations,
