+++ src/argilla/server/apis/v0/validators/text_classification.py
@@ -17,9 +17,9 @@ from typing import List, Optional, Set, Type
 from fastapi import Depends
 
 from argilla.server.apis.v0.models.dataset_settings import TextClassificationSettings
-from argilla.server.apis.v0.models.datasets import Dataset
 from argilla.server.commons.models import TaskType
 from argilla.server.errors import BadRequestError, EntityNotFoundError
+from argilla.server.schemas.datasets import Dataset
 from argilla.server.security.model import User
 from argilla.server.services.datasets import DatasetsService, ServiceBaseDatasetSettings
 from argilla.server.services.tasks.text_classification.metrics import DatasetLabels
